+++ espflash/src/bin/espflash.rs
@@ -16,6 +16,7 @@ use espflash::{
 };
 use log::{LevelFilter, debug, info};
 use miette::{IntoDiagnostic, Result, WrapErr};
+use object::{File, Object, ObjectSymbol};
 
 #[derive(Debug, Parser)]
 #[command(about, max_term_width = 100, propagate_version = true, version)]
@@ -234,6 +235,16 @@ fn flash(args: FlashArgs, config: &Config) -> Result<()> {
 
     // Read the ELF data from the build path and load it to the target.
     let elf_data = fs::read(&args.image).into_diagnostic()?;
+
+    if args.flash_args.image.check_app_descriptor.unwrap_or(true) {
+        let object = File::parse(elf_data.as_slice()).into_diagnostic()?;
+        let section = object.section_by_name(".rodata_desc").is_some();
+        let symbol = object.symbols().any(|sym| sym.name() == Ok("esp_app_desc"));
+
+        if !section || !symbol {
+            return Err(Error::AppDescriptorNotPresent).into_diagnostic();
+        }
+    }
 
     print_board_info(&mut flasher)?;
     ensure_chip_compatibility(chip, Some(elf_data.as_slice()))?;